Add ICU & improvement
