+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * Add phonex algorithm
 * Rework for Tantivy 0.20
 * Remove deprecated trim filter
+* Remove some features
+* Change how to construct a `ICUNormalizer2TokenFilter`
 
 ## 0.8.0
