+++ src/commons/limit/token_filter.rs
@@ -57,7 +57,7 @@ impl From<usize> for LimitTokenCountFilter {
 impl TokenFilter for LimitTokenCountFilter {
     type Tokenizer<T: Tokenizer> = LimitTokenCountFilterWrapper<T>;
 
-    fn transform<T: Tokenizer>(self, token_stream: T) -> LimitTokenCountFilterWrapper<T> {
+    fn transform<T: Tokenizer>(self, token_stream: T) -> Self::Tokenizer<T> {
         LimitTokenCountFilterWrapper {
             inner: token_stream,
             count: self.max_tokens,
