+++ src/commons/reverse/token_filter.rs
@@ -28,7 +28,7 @@ pub struct ReverseTokenFilter;
 impl TokenFilter for ReverseTokenFilter {
     type Tokenizer<T: Tokenizer> = ReverseFilterWrapper<T>;
 
-    fn transform<T: Tokenizer>(self, token_stream: T) -> ReverseFilterWrapper<T> {
+    fn transform<T: Tokenizer>(self, token_stream: T) -> Self::Tokenizer<T> {
         ReverseFilterWrapper {
             inner: token_stream,
         }
