+++ src/icu/tokenizer/breaking_rules/MyanmarSyllable.rbbi
